chien updat2

+++ QuanLyThuVien2/QuanLyThuVien2/UpdatesBook.cs
@@ -29,118 +29,127 @@ namespace QuanLyThuVien2
             return false;
         }
 
+     
+
         public int numberUndo = 0;
-      
+
         string masach;
         private void Sua_Click(object sender, EventArgs e)
         {
-
-            if (txtMASACH.Text != "")
+            if (txtMASACH.Text == "")
             {
-                if (txtTENSACH.Text != "")
+                MessageBox.Show("Bạn hãy chọn 1 hàng dữ liệu để sửa");
+            }
+            else
+            {
+                if (txtMASACH.Text != "")
+                {
+                    if (txtTENSACH.Text != "")
 
-                {              
-                      if (txtNAMXB.Text != "")
+                    {
+                        if (txtNAMXB.Text != "")
                         {
-                        if (!hasSpecialChar(txtNAMXB.Text))
-                        {
-                            MessageBox.Show("Năm xuất bản chỉ được chứa ký tự số");
-                        }
-                        else
-                        {
-
-                            if (txtSOTRANG.Text != "")
+                            if (!hasSpecialChar(txtNAMXB.Text))
                             {
-                                if (!hasSpecialChar(txtSOTRANG.Text))
+                                MessageBox.Show("Năm xuất bản chỉ được chứa ký tự số");
+                            }
+                            else
+                            {
+
+                                if (txtSOTRANG.Text != "")
                                 {
-                                    MessageBox.Show("Số trang chỉ được chứa ký tự số");
+                                    if (!hasSpecialChar(txtSOTRANG.Text))
+                                    {
+                                        MessageBox.Show("Số trang chỉ được chứa ký tự số");
+
+                                    }
+                                    else
+                                    {
+                                        if (txtSOLUONG.Text != "")
+                                        {
+                                            if (!hasSpecialChar(txtSOLUONG.Text))
+                                            {
+                                                MessageBox.Show("Số lượng chỉ được chứa ký tự số");
+                                            }
+                                            else
+                                            {
+                                                if (txtsachhong.Text != "")
+                                                {
+                                                    if (!hasSpecialChar(txtsachhong.Text))
+                                                    {
+                                                        MessageBox.Show("Số sách hỏng chỉ được chứa ký tự số");
+                                                    }
+                                                    else
+                                                    {
+                                                        if (maskedTextBox1.Text != "")
+                                                        {
+
+                                                            try
+                                                            {
+                                                                string strUpdate = "Update tblSach set MASACH='" + txtMASACH.Text + "',TENSACH= N'" + txtTENSACH.Text + "',MATG='" + cboMATG.Text + "',MANXB='" + cboMANXB.Text + "',MaLv='" + cboMALv.Text + "',NAMXB='" + txtNAMXB.Text + "',SOTRANG='" + txtSOTRANG.Text + "',SOLUONG='" + txtSOLUONG.Text + "',SOSACHHONG='" + txtsachhong.Text + "',NGAYNHAP='" + maskedTextBox1.Text + "',GHICHU='" + txtGHICHU.Text + "' where MASACH='" + masach + "'";
+                                                                cls.ThucThiSQLTheoPKN(strUpdate);
+                                                                cls.LoadData2DataGridView(dataGridView1, "select * from tblSach");
+                                                                MessageBox.Show("Sửa thành công");
+                                                                txtMASACH.Text = "";
+                                                                txtTENSACH.Text = "";
+                                                                txtNAMXB.Text = "";
+                                                                txtSOTRANG.Text = "";
+                                                                txtSOLUONG.Text = "";
+                                                                txtsachhong.Text = "";
+                                                                maskedTextBox1.Text = "";
+                                                                txtGHICHU.Text = "";
+                                                                cboMALv.Text = "";
+                                                                cbotenLV.Text = "";
+                                                                cboMATG.Text = "";
+                                                                cbotenTG.Text = "";
+                                                                cboMANXB.Text = "";
+                                                                cbotenNXB.Text = "";
+                                                                numberUndo = 0;
+
+                                                            }
+                                                            catch { }
+                                                        }
+                                                        else { MessageBox.Show("Ngày nhập không được để trống"); }
+                                                    }
+                                                }
+                                                else
+                                                {
+                                                    MessageBox.Show("Số sách hỏng không được để trống");
+                                                }
+                                            }
+                                        }
+                                        else
+                                        {
+                                            MessageBox.Show("Số lượng sách không được bỏ trống");
+                                        }
+                                    }
 
                                 }
                                 else
                                 {
-                                    if (txtSOLUONG.Text != "")
-                                    {
-                                        if (!hasSpecialChar(txtSOLUONG.Text))
-                                        {
-                                            MessageBox.Show("Số lượng chỉ được chứa ký tự số");
-                                        }
-                                        else
-                                        {
-                                            if (txtsachhong.Text != "")
-                                            {
-                                                if (!hasSpecialChar(txtsachhong.Text))
-                                                {
-                                                    MessageBox.Show("Số sách hỏng chỉ được chứa ký tự số");
-                                                }
-                                                else
-                                                {
-                                                    if (maskedTextBox1.Text != "")
-                                                    {
-
-                                                        try
-                                                        {
-                                                            string strUpdate = "Update tblSach set MASACH='" + txtMASACH.Text + "',TENSACH= N'" + txtTENSACH.Text + "',MATG='" + cboMATG.Text + "',MANXB='" + cboMANXB.Text + "',MaLv='" + cboMALv.Text + "',NAMXB='" + txtNAMXB.Text + "',SOTRANG='" + txtSOTRANG.Text + "',SOLUONG='" + txtSOLUONG.Text + "',SOSACHHONG='" + txtsachhong.Text + "',NGAYNHAP='" + maskedTextBox1.Text + "',GHICHU='" + txtGHICHU.Text + "' where MASACH='" + masach + "'";
-                                                            cls.ThucThiSQLTheoPKN(strUpdate);
-                                                            cls.LoadData2DataGridView(dataGridView1, "select * from tblSach");
-                                                            MessageBox.Show("Sửa thành công");
-                                                            txtMASACH.Text = "";
-                                                            txtTENSACH.Text = "";
-                                                            txtNAMXB.Text = "";
-                                                            txtSOTRANG.Text = "";
-                                                            txtSOLUONG.Text = "";
-                                                            txtsachhong.Text = "";
-                                                            maskedTextBox1.Text = "";
-                                                            txtGHICHU.Text = "";
-                                                            cboMALv.Text = "";
-                                                            cbotenLV.Text = "";
-                                                            cboMATG.Text = "";
-                                                            cbotenTG.Text = "";
-                                                            cboMANXB.Text = "";
-                                                            cbotenNXB.Text = "";
-                                                            numberUndo = 0;
-
-                                                        }
-                                                        catch { }
-                                                    }
-                                                    else { MessageBox.Show("Ngày nhập không được để trống"); }
-                                                }
-                                            }
-                                            else
-                                            {
-                                                MessageBox.Show("Số sách hỏng không được để trống");
-                                            }
-                                        }
-                                    }
-                                    else
-                                    {
-                                        MessageBox.Show("Số lượng sách không được bỏ trống");
-                                    }
+                                    MessageBox.Show("Số trang không được để trống");
                                 }
-
                             }
-                            else
-                            {
-                                MessageBox.Show("Số trang không được để trống");
-                            }
-                        }
                         }
                         else { MessageBox.Show("Năm xuất bản không được để trống"); }
 
-                    
+
+                    }
+                    else
+                    {
+                        MessageBox.Show("Tên sách không được để trống");
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("Tên sách không được để trống");
+                    MessageBox.Show("Mã sách không được để trống ");
                 }
             }
-            else
-            {
-                MessageBox.Show("Mã sách không được để trống ");
-            }
+            
 
         }
 
-       
+
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             try
@@ -156,6 +165,9 @@ namespace QuanLyThuVien2
                 txtsachhong.Text = dataGridView1.Rows[e.RowIndex].Cells[8].Value.ToString();
                 maskedTextBox1.Text = dataGridView1.Rows[e.RowIndex].Cells[9].Value.ToString();
                 txtGHICHU.Text = dataGridView1.Rows[e.RowIndex].Cells[10].Value.ToString();
+      
+
+
                 masach = txtMASACH.Text;
 
             }
@@ -323,44 +335,50 @@ namespace QuanLyThuVien2
 
         private void Xoa_Click(object sender, EventArgs e)
         {
-            try
+            if (txtMASACH.Text == "")
             {
-                if (MessageBox.Show("Bạn có muốn xóa không?(Y/N)", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
-                {
-                    undoMS = txtMASACH.Text;
-                    undoTS = txtTENSACH.Text;
-                    undoMaLV = cboMALv.Text;
-                    undoNXB = txtNAMXB.Text;
-                    undoST = txtSOTRANG.Text;
-                    undoSL = txtSOLUONG.Text;
-                    undoSSH = txtsachhong.Text;
-                    undoGHICHU = txtGHICHU.Text;
-                    undoNN = maskedTextBox1.Text;
-                    undoMaNXB = cboMANXB.Text;
-                    undoMaTG = cboMATG.Text;
-                    string strDelete = "Delete from tblSach where MASACH='" + txtMASACH.Text + "'";
-                    cls.ThucThiSQLTheoKetNoi(strDelete);
-                    cls.LoadData2DataGridView(dataGridView1, "select * from tblSach");
-                    MessageBox.Show("Xóa thành công !!!");
-                    txtMASACH.Text = "";
-                    txtTENSACH.Text = "";
-                    txtNAMXB.Text = "";
-                    txtSOTRANG.Text = "";
-                    txtSOLUONG.Text = "";
-                    txtsachhong.Text = "";
-                    maskedTextBox1.Text = "";
-                    txtGHICHU.Text = "";
-                    cboMALv.Text = "";
-                    cbotenLV.Text = "";
-                    cboMATG.Text = "";
-                    cbotenTG.Text = "";
-                    cboMANXB.Text = "";
-                    cbotenNXB.Text = "";
-                    numberUndo = 1;
-                }
+                MessageBox.Show("Bạn hãy chọn 1 dòng để xóa");
             }
-            catch { }
-             
+            else
+            {
+                try
+                {
+                    if (MessageBox.Show("Bạn có muốn xóa không?(Y/N)", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                    {
+                        undoMS = txtMASACH.Text;
+                        undoTS = txtTENSACH.Text;
+                        undoMaLV = cboMALv.Text;
+                        undoNXB = txtNAMXB.Text;
+                        undoST = txtSOTRANG.Text;
+                        undoSL = txtSOLUONG.Text;
+                        undoSSH = txtsachhong.Text;
+                        undoGHICHU = txtGHICHU.Text;
+                        undoNN = maskedTextBox1.Text;
+                        undoMaNXB = cboMANXB.Text;
+                        undoMaTG = cboMATG.Text;
+                        string strDelete = "Delete from tblSach where MASACH='" + txtMASACH.Text + "'";
+                        cls.ThucThiSQLTheoKetNoi(strDelete);
+                        cls.LoadData2DataGridView(dataGridView1, "select * from tblSach");
+                        MessageBox.Show("Xóa thành công !!!");
+                        txtMASACH.Text = "";
+                        txtTENSACH.Text = "";
+                        txtNAMXB.Text = "";
+                        txtSOTRANG.Text = "";
+                        txtSOLUONG.Text = "";
+                        txtsachhong.Text = "";
+                        maskedTextBox1.Text = "";
+                        txtGHICHU.Text = "";
+                        cboMALv.Text = "";
+                        cbotenLV.Text = "";
+                        cboMATG.Text = "";
+                        cbotenTG.Text = "";
+                        cboMANXB.Text = "";
+                        cbotenNXB.Text = "";
+                        numberUndo = 1;
+                    }
+                }
+                catch { MessageBox.Show("Phải xóa thông tin liên quan đến cuốn sách này trước"); }
+            }
         }
 
         private void Thoat_Click(object sender, EventArgs e)
@@ -368,6 +386,13 @@ namespace QuanLyThuVien2
             Close();
         }
         public string undoMS, undoTS, undoMaLV, undoNXB, undoST, undoSL, undoSSH, undoGHICHU, undoNN, undoMaNXB, undoMaTG;
+
+        private void btSearch_Click(object sender, EventArgs e)
+        {
+            SearchBooks sb = new SearchBooks();
+            sb.Show();
+        }
+
         private void Undo_Click(object sender, EventArgs e)
         {
             if (numberUndo == 1)
